+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 # 0.1.1 (Unreleased)
 
+- Fix `alias` should not be considered objects but only used as types [[GH-76](https://github.com/ferranbt/fastssz/issues/76)]
 - Fix the exclude of types from generation if they are set with the `exclude-objs` flag [[GH-76](https://github.com/ferranbt/fastssz/issues/76)]
 - Add `version` command to `sszgen` [[GH-74](https://github.com/ferranbt/fastssz/issues/74)]
 - Support `bellatrix`, `altair` and `phase0` forks in spec tests command to `sszgen` [[GH-73](https://github.com/ferranbt/fastssz/issues/73)]